2024/07/23 |

+++ recbole/evaluator/metrics.py
@@ -55,7 +55,7 @@ def mrr_(pos_index, pos_len):
 
     """
     idxs = pos_index.argmax(axis=1)
-    result = np.zeros_like(pos_index, dtype=np.float)
+    result = np.zeros_like(pos_index, dtype=np.cfloat)
     for row, idx in enumerate(idxs):
         if pos_index[row, idx] > 0:
             result[row, idx:] = 1 / (idx + 1)
